fix: extract fish shell SSH support to pass file-length budget
Keep the fish-shell SSH invariant in the existing CLI command builder while moving the wrapper helpers and regression coverage out of files that exceeded CI's line-count budget.

Constraint: Do not raise .github/swift-file-length-budget.tsv

Constraint: Preserve POSIX /bin/sh wrapping for OpenSSH LocalCommand and RemoteCommand so fish login shells parse the command line

Rejected: Bump file-length budget | CI failure explicitly requires extraction instead

Confidence: high

Scope-risk: narrow

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: swiftc -typecheck CLI/cmux.swift CLI/SocketOperationTelemetry.swift CLI/cmux_open.swift CLI/CMUXCLI+MoveTabToNewWorkspace.swift CLI/CMUXCLI+DocsSettings.swift CLI/CMUXCLI+ThemeSupport.swift CLI/CMUXCLI+Themes.swift CLI/CMUXCLI+TopRendering.swift CLI/CMUXFishShellSupport.swift Sources/RemoteRelayZshBootstrap.swift

Not-tested: Local XCTest run, per repository testing policy
